* Fixes as per Thien's suggestions - restored RestRemoteServer.cpp' * Changes to RemoteSimulatorQPU
